Changes done to admit args from main, WeatherController revised, OpenWeatherMapProvider has been finished?, SQLiteWeatherStore finally stores data correctly (still waiting for update / insert check when new data provided)

+++ src/main/java/dacd/riveromonzon/practice1/model/Location.java
@@ -15,11 +15,11 @@ public class Location {
 		return name;
 	}
 
-	public double getLatitude() {
+	public Float getLatitude() {
 		return latitude;
 	}
 
-	public double getLongitude() {
+	public Float getLongitude() {
 		return longitude;
 	}
 }
